+++ lib/torchvision/transforms/center_crop.rb
@@ -1,0 +1,13 @@
+module TorchVision
+  module Transforms
+    class CenterCrop < Torch::NN::Module
+      def initialize(size)
+        @size = size
+      end
+
+      def forward(img)
+        F.center_crop(img, @size)
+      end
+    end
+  end
+end