+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.1 (unreleased)
 
+- Added `CenterCrop` transform
 - Added `crop` method
 
 ## 0.2.0 (2021-03-11)